Support CafeDelites recipes
This will scrape the page (with the recipe) and return a Recipe struct
with the ingredients from the page.

Some work is needed to "clean" the ingredient names (i.e. "Skinless
Chicken Thighs" instead of "2 pounds (1 kg) boneless, skinless chicken
thighs") and to identify the type of the Ingredient (at the moment all
are marked as "N/A").